requirements.txt
python-telegram-bot==21.8
openai>=1.42.0
pydantic>=2.8.2
python-dotenv>=1.0.1
PyMuPDF>=1.24.10